+++ healthbarview/src/main/java/org/ogasimli/healthbarview/MinValue.java
@@ -133,7 +133,7 @@ class MinValue extends BaseValue {
     }
 
     @Override
-    void setValue(float value) {
+    void setValue(double value) {
         super.setValue(value);
         mView.invalidate();
     }
